🎨 Combine private and public endpoints in favor of UnauthorizedException

+++ app/Controllers/Http/v1/OrganisationController.ts
@@ -1,6 +1,7 @@
 import { HttpContextContract } from '@ioc:Adonis/Core/HttpContext';
 import OrganisationValidator from 'App/Validators/v1/Private/OrganisationValidator';
 import Organisation from 'App/Models/Organisation';
+import UnauthorizedException from 'App/Exceptions/UnauthorizedException';
 
 // TODO(matthiasrohmer): Add permissions
 export default class OrganisationController {
@@ -14,6 +15,10 @@ export default class OrganisationController {
   }
 
   public async store({ request, response, auth }: HttpContextContract) {
+    if (!auth.user) {
+      throw new UnauthorizedException();
+    }
+
     const data = await request.validate(OrganisationValidator);
     const organisation = await Organisation.create(data);
 
@@ -26,9 +31,11 @@ export default class OrganisationController {
     });
   }
 
-  public async show({ request, response, params }: HttpContextContract) {
+  public async show({ request, response, params, auth }: HttpContextContract) {
     const organisation = await Organisation.findOrFail(params.id);
-    await organisation.load('members');
+    if (auth.user) {
+      await organisation.load('members');
+    }
 
     return response.ok({
       organisation,
@@ -36,7 +43,11 @@ export default class OrganisationController {
     });
   }
 
-  public async update({ request, response, params }: HttpContextContract) {
+  public async update({ request, response, params, auth }: HttpContextContract) {
+    if (!auth.user) {
+      throw new UnauthorizedException();
+    }
+
     const data = await request.validate(OrganisationValidator);
 
     const organisation = await Organisation.findOrFail(params.id);
@@ -50,7 +61,11 @@ export default class OrganisationController {
     });
   }
 
-  public async destroy({ request, response, params }: HttpContextContract) {
+  public async destroy({ request, response, params, auth }: HttpContextContract) {
+    if (!auth.user) {
+      throw new UnauthorizedException();
+    }
+
     const organisation = await Organisation.findOrFail(params.id);
     await organisation.delete();
